requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
pydantic[email]
python-multipart==0.0.20
pyjwt==2.10.1
passlib[bcrypt]
pwdlib[argon2]
sqlmodel==0.0.27
python-dotenv==1.2.1
